+++ examples/island_cluster_anywhere_in_gcp_design/main.tf
@@ -27,7 +27,7 @@ resource "google_service_account" "gke-sa" {
 
 module "gke" {
   source  = "terraform-google-modules/kubernetes-engine/google//modules/beta-private-cluster"
-  version = "~> 34.0"
+  version = "~> 36.0"
 
   for_each = { for k, v in var.gke_spokes : k => v }
 
@@ -65,7 +65,6 @@ module "gke" {
       disk_size_gb              = 100
       disk_type                 = "pd-standard"
       image_type                = "COS_CONTAINERD"
-      logging_variant           = "DEFAULT"
       auto_repair               = true
       auto_upgrade              = true
       service_account           = google_service_account.gke-sa[each.key].email