+++ examples/safer_cluster_iap_bastion/cluster.tf
@@ -16,7 +16,7 @@
 
 module "gke" {
   source  = "terraform-google-modules/kubernetes-engine/google//modules/safer-cluster"
-  version = "~> 34.0"
+  version = "~> 36.0"
 
   project_id              = module.enabled_google_apis.project_id
   name                    = var.cluster_name
@@ -25,7 +25,7 @@ module "gke" {
   subnetwork              = module.vpc.subnets_names[0]
   ip_range_pods           = module.vpc.subnets_secondary_ranges[0][0].range_name
   ip_range_services       = module.vpc.subnets_secondary_ranges[0][1].range_name
-  enable_private_endpoint = false
+  enable_private_endpoint = true
   deletion_protection     = false
   master_authorized_networks = [{
     cidr_block   = "${module.bastion.ip_address}/32"