+++ test/fixtures/node_pool/outputs.tf
@@ -60,8 +60,7 @@ output "master_kubernetes_version" {
 }
 
 output "kubernetes_endpoint" {
-  sensitive = true
-  value     = module.example.kubernetes_endpoint
+  value = module.example.kubernetes_endpoint
 }
 
 output "client_token" {
@@ -82,4 +81,12 @@ output "service_account" {
 
 output "registry_project_ids" {
   value = var.registry_project_ids
+}
+
+output "random_string" {
+  value = random_string.suffix.result
+}
+
+output "compute_engine_service_account" {
+  value = local.compute_engine_service_account
 }