+++ test/fixtures/simple_windows_node_pool/outputs.tf
@@ -42,8 +42,7 @@ output "master_kubernetes_version" {
 }
 
 output "kubernetes_endpoint" {
-  sensitive = true
-  value     = module.this.kubernetes_endpoint
+  value = module.this.kubernetes_endpoint
 }
 
 output "client_token" {