+++ test/integration/testutils/utils.go
@@ -1,4 +1,4 @@
-// Copyright 2022-2024 Google LLC
+// Copyright 2022-2025 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,14 +15,18 @@
 package testutils
 
 import (
+	"fmt"
 	"slices"
 	"strings"
 	"testing"
 	"time"
 
+	"github.com/GoogleCloudPlatform/cloud-foundation-toolkit/infra/blueprint-test/pkg/golden"
 	"github.com/GoogleCloudPlatform/cloud-foundation-toolkit/infra/blueprint-test/pkg/tft"
+	"github.com/GoogleCloudPlatform/cloud-foundation-toolkit/infra/blueprint-test/pkg/utils"
 	tfjson "github.com/hashicorp/terraform-json"
 	"github.com/stretchr/testify/assert"
+	"github.com/tidwall/gjson"
 )
 
 var (
@@ -35,7 +39,12 @@ var (
 
 		// API Rate limit exceeded errors can be retried.
 		".*rateLimitExceeded.*": "Rate limit exceeded.",
+
+		// Internal errors can be retried
+		".*Error code 13, message: an internal error has occurred": "Internal error.",
 	}
+
+	ClusterAlwaysExemptPaths = []string{"nodePools"} // node pools are separately checked by name
 )
 
 func GetTestProjectFromSetup(t *testing.T, idx int) string {
@@ -65,5 +74,38 @@ func TGKEVerifyExemptResources(t *testing.T, b *tft.TFBlueprintTest, assert *ass
 			continue
 		}
 		assert.Equal(tfjson.Actions{tfjson.ActionNoop}, r.Change.Actions, "Plan must be no-op for resource: %s", r.Address)
+	}
+}
+
+// TGKEAssertGolden asserts a cluster and listed node pools against paths in golden image
+func TGKEAssertGolden(assert *assert.Assertions, golden *golden.GoldenFile, clusterJson *gjson.Result, nodePools []string, exemptClusterPaths []string) {
+	// Retrieve golden paths
+	clusterCheckPaths := utils.GetTerminalJSONPaths(golden.GetJSON())
+
+	// Remove exempt cluster paths
+	exemptPaths := slices.Concat(exemptClusterPaths, ClusterAlwaysExemptPaths)
+	clusterCheckPaths = slices.DeleteFunc(clusterCheckPaths, func(s string) bool {
+		for _, exempPath := range exemptPaths {
+			if strings.HasPrefix(s, exempPath) {
+				return true
+			}
+		}
+		return false
+	})
+
+	// Cluster assertions
+	golden.JSONPathEqs(assert, *clusterJson, clusterCheckPaths)
+
+	// NodePool assertions
+	for _, nodePool := range nodePools {
+		assert.Truef(clusterJson.Get(fmt.Sprintf("nodePools.#(name==%s).name", nodePool)).Exists(), "NodePool not found: %s", nodePool)
+
+		nodeCheckPaths := utils.GetTerminalJSONPaths(golden.GetJSON().Get(fmt.Sprintf("nodePools.#(name==%s)", nodePool)))
+
+		for _, nodeCheckPath := range nodeCheckPaths {
+			gotData := golden.ApplySanitizers(clusterJson.Get(fmt.Sprintf("nodePools.#(name==%s)", nodePool)).Get(nodeCheckPath).String())
+			gfData := golden.GetJSON().Get(fmt.Sprintf("nodePools.#(name==%s)", nodePool)).Get(nodeCheckPath).String()
+			assert.Equalf(gfData, gotData, "For node %q path %q expected %q to match fixture %q", nodePool, nodeCheckPath, gotData, gfData)
+		}
 	}
 }