+++ test/setup/main.tf
@@ -39,13 +39,14 @@ locals {
     "iamcredentials.googleapis.com",
     "gkeconnect.googleapis.com",
     "privateca.googleapis.com",
-    "gkehub.googleapis.com"
+    "gkehub.googleapis.com",
+    "cloudasset.googleapis.com"
   ]
 }
 
 module "gke-project-1" {
   source  = "terraform-google-modules/project-factory/google"
-  version = "~> 17.0"
+  version = "~> 18.0"
 
   name                     = "ci-gke-${random_id.random_project_id_suffix.hex}"
   random_project_id        = true
@@ -69,7 +70,7 @@ module "gke-project-1" {
 
 module "gke-project-2" {
   source  = "terraform-google-modules/project-factory/google"
-  version = "~> 17.0"
+  version = "~> 18.0"
 
   name              = "ci-gke-${random_id.random_project_id_suffix.hex}"
   random_project_id = true
@@ -91,7 +92,7 @@ module "gke-project-2" {
 # apis as documented https://cloud.google.com/service-mesh/docs/scripted-install/reference#setting_up_your_project
 module "gke-project-asm" {
   source  = "terraform-google-modules/project-factory/google"
-  version = "~> 17.0"
+  version = "~> 18.0"
 
   name              = "ci-gke-asm-${random_id.random_project_id_suffix.hex}"
   random_project_id = true
@@ -106,7 +107,7 @@ module "gke-project-asm" {
 
 module "gke-project-fleet" {
   source  = "terraform-google-modules/project-factory/google"
-  version = "~> 17.0"
+  version = "~> 18.0"
 
   name              = "ci-gke-fleet-${random_id.random_project_id_suffix.hex}"
   random_project_id = true